+++ MemoryPool/recibir.h
@@ -29,7 +29,7 @@ typedef struct t_paquete_select
 {
 	char* nombre_tabla;
 	uint32_t nombre_tabla_long;
-	int valor_key;
+	uint16_t valor_key;
 }t_paquete_select;
 
 typedef struct t_paquete_insert
@@ -37,9 +37,9 @@ typedef struct t_paquete_insert
 	uint32_t nombre_tabla_long; //Longitud del nombre de la tabla
 	uint32_t value_long; 		//Longitud del value
 	char* nombre_tabla;
-	int valor_key;
+	uint16_t valor_key;
 	char* value;
-	uint16_t timestamp;
+	int timestamp;
 }t_paquete_insert;
 
 //--------------------RECIBIR PAQUETE
